Reconcile duplicate PosixFile and LinuxPosixFile
- Converted PosixFile to open class.
- Replaced LinuxPosixFile with thin subclass of PosixFile.
- Moved buffered writeLines from LinuxPosixFile into PosixFile.
- Retained Linux-specific companion object methods in LinuxPosixFile.
